Remove conflicting command-line flags
We mostly did a good job of picking flags that CPython doesn't use (or at least, that
its manpage doesn't advertise), but there are a couple that need to be changed.

One is for TRAP which is a feature I don't even remember anymore, so that's removed.

The other is for stats -- '-s' is now changed to '-T'.  This will be annoying to change
on the perf-tracking side.