Add a bunch of missing symlinks
We symlink the CPython tests into our tests directory to control which
ones we will attempt to run.  A bunch of the tests though need other
test files in order to run, and we hadn't symlinked those in.

At some point we should switch to symlinking the Lib/test directory instead
of individual files, but for now this helps.