Split test_lib2to3 into small and large variants
and only run the small variant in debug mode to avoid timeouts.
Previously I had tried splitting it in half, but it looks like that's
both not enough to avoid timeouts, and it also has a race condition when
trying to run two copies of the test at the same time (one will read the
not-fully-written pickled output of the other).