Remove this failing test
It was previously incorrectly succeeding: we happened to throw a
similar-enough exception that the test thought we were testing the right
thing.  Once we fixed that bug, it turned up that we didn't throw the
exception that the test is expecting.

I don't like removing tests but this isn't any worse than the current
situation of not really testing this functionality.