Change the field name vreg_name to index_id in BST_ImportFrom and BST_ImportName
This commit is a preperation of the commit which for add co_names
field.  BST_StoreName and some other nodes use `index_id` and
InteredString to store variable name. But BST_ImportName use vreg
and vreg_name. We can't get the module name inside code object by
vreg. So change it to the way which BST_StoreName used.